[#3636] section5 indicator.periods
